+++ lizard_damage/tasks.py
@@ -191,13 +191,13 @@ def calculate_damage(damage_scenario_id, username=None, taskname=None, loglevel=
 
     if errors == 0:
         logger.info("creating email task")
-        subject = 'Schademodule: Resultaten beschikbaar voor scenario %s ' % damage_scenario.name
+        subject = 'STOWA Schade Calculator: Resultaten beschikbaar voor scenario %s ' % damage_scenario.name
         send_email_to_task(damage_scenario.id, 'email_ready', subject, username=username)
         logger.info("finished")
     else:
         logger.info("there were errors")
         logger.info("creating email task for error")
-        subject = 'Schademodule: scenario %s heeft fouten' % damage_scenario.name
+        subject = 'STOWA Schade Calculator: scenario %s heeft fouten' % damage_scenario.name
         send_email_to_task(damage_scenario.id, 'email_error', subject, username=username,
                            email='jack.ha@nelen-schuurmans.nl')
         logger.info("finished with errors")